+++ tt_eager/tt_dnn/op_library/bcast/bcast_op.cpp
@@ -145,13 +145,15 @@ operation::ProgramWithCallbacks EltwiseBinaryBroadcast::create_program(const std
 const operation::Hash EltwiseBinaryBroadcast::compute_program_hash(
     const std::vector<Tensor> &input_tensors) const {
     auto parallelization_strategy = this->get_parallelization_strategy(input_tensors);
+    bool bcast_scalar = (input_tensors.at(1).shape()[-2] * input_tensors.at(1).shape()[-1] == 1) && this->dim == BcastOpDim::HW;
     return operation::hash_operation<EltwiseBinaryBroadcast>(
         *this,
         parallelization_strategy,
         input_tensors.at(0).memory_config(),
         input_tensors.at(0).dtype(),
         input_tensors.at(1).memory_config(),
-        input_tensors.at(1).dtype());
+        input_tensors.at(1).dtype(),
+        bcast_scalar);
 }
 
 BcastOpParallelizationStrategy EltwiseBinaryBroadcast::get_parallelization_strategy(const std::vector<Tensor> &input_tensors) const {
