+++ tt_eager/tt_dnn/op_library/layernorm/kernels/dataflow/reader_mcast_receiver_unary_sharded_ln.cpp
@@ -17,11 +17,16 @@ void kernel_main() {
     constexpr uint32_t num_all_to_all_workers          = get_compile_time_arg_val(5);
     constexpr uint32_t num_tiles_per_worker            = get_compile_time_arg_val(6);
     constexpr uint32_t num_tiles_per_worker_last       = get_compile_time_arg_val(7);
+    constexpr bool row_major                           = (bool) get_compile_time_arg_val(8);
+    constexpr uint32_t num_x                           = get_compile_time_arg_val(9);
+    constexpr uint32_t num_y                           = get_compile_time_arg_val(10);
 
     const bool is_last_all_to_all_worker                = get_arg_val<uint32_t>(0);
     const uint32_t all_to_all_tile_offset_bytes         = get_arg_val<uint32_t>(1);
-    const uint32_t noc_same_coord                       = get_arg_val<uint32_t>(2);
-    volatile tt_l1_ptr uint32_t * noc_diff_coord        = (volatile tt_l1_ptr uint32_t*)(get_arg_addr(3));
+    const uint32_t start_x                              = get_arg_val<uint32_t>(2);
+    const uint32_t start_y                              = get_arg_val<uint32_t>(3);
+    volatile tt_l1_ptr uint32_t * in0_remote_noc_x          = (volatile tt_l1_ptr uint32_t*)(get_arg_addr(4));
+    volatile tt_l1_ptr uint32_t * in0_remote_noc_y          = (volatile tt_l1_ptr uint32_t*)(get_arg_addr(4 + num_x));
 
     const uint32_t num_tiles_to_read = is_last_all_to_all_worker ? num_tiles_per_worker_last : num_tiles_per_worker;
 
@@ -37,41 +42,72 @@ void kernel_main() {
     const uint32_t single_tile_size_bytes = get_tile_size(cb_ex_partial); // tile size
     const DataFormat data_format = get_dataformat(cb_ex_partial); // data format
 
+    uint64_t remote_noc_addrs[is_all_to_all_worker ? num_blocks : 1];
+    if constexpr (is_all_to_all_worker) {
+        uint32_t x = start_x, y = start_y;
+        for (uint32_t i = 0; i < num_blocks; ++i) {
+            remote_noc_addrs[i] = get_noc_addr(in0_remote_noc_x[x], in0_remote_noc_y[y], 0);
+            if constexpr(row_major) {
+                ++y;
+                if (y == num_y) {
+                    y = 0;
+                    ++x;
+                    if (x == num_x) {
+                        x = 0;
+                    }
+                }
+            } else {
+                ++x;
+                if (x == num_x) {
+                    x = 0;
+                    ++y;
+                    if (y == num_y) {
+                        y = 0;
+                    }
+                }
+            }
+        }
+    } else {
+        remote_noc_addrs[0] = get_noc_addr(in0_remote_noc_x[0], in0_remote_noc_y[0], 0);
+    }
+
     volatile tt_l1_ptr uint32_t* reduce_receiver_semaphore_addr_ptr = reinterpret_cast<volatile tt_l1_ptr uint32_t*>(reduce_receiver_semaphore_addr);
     volatile tt_l1_ptr uint32_t* reduce_sender_semaphore_addr_ptr = reinterpret_cast<volatile tt_l1_ptr uint32_t*>(reduce_sender_semaphore_addr);
 
-    // global reduce
-    if constexpr(is_all_to_all_worker) {
+    const uint64_t reduce_receiver_semaphore_noc_addr = get_noc_addr(in0_remote_noc_x[0], in0_remote_noc_y[0], reduce_receiver_semaphore_addr);
+
+    const auto& global_reduce_receiver = [&](const uint32_t cb_partial, const uint32_t cb_external, const uint32_t cb_ex, const uint32_t cb_ex_global) __attribute__((always_inline))
+    {
+        // global reduce
         // wait for local data ready
-        cb_wait_front(cb_ex_partial, block_h);
+        cb_wait_front(cb_partial, block_h);
 
         // inc top core
         noc_semaphore_set(reduce_sender_semaphore_addr_ptr, INVALID);
-        const uint64_t reduce_receiver_semaphore_noc_addr = get_noc_addr(noc_same_coord, 1, reduce_receiver_semaphore_addr);
         noc_semaphore_inc(reduce_receiver_semaphore_noc_addr, 1);
         noc_semaphore_wait(reduce_sender_semaphore_addr_ptr, VALID);
 
-        // read data from other cores
-        uint32_t l1_read_addr_ex_par = get_read_ptr(cb_ex_partial);
-        l1_read_addr_ex_par += all_to_all_tile_offset_bytes;
-        for (uint32_t i = 0; i < num_tiles_to_read; i++) {
-            uint32_t l1_write_addr_external = get_write_ptr(cb_ex_external);
-            for(uint32_t block = 0; block < num_blocks; block++) {
-                cb_reserve_back(cb_ex_external, 1);
-                uint64_t noc_addr_ex_par = get_noc_addr(noc_same_coord, noc_diff_coord[block], l1_read_addr_ex_par);
-                noc_async_read_one_packet(noc_addr_ex_par, l1_write_addr_external, single_tile_size_bytes);
-                l1_write_addr_external += single_tile_size_bytes;
+        if constexpr (is_all_to_all_worker) {
+            // read data from other cores
+            uint32_t l1_read_addr_ex_par = get_read_ptr(cb_partial);
+            l1_read_addr_ex_par += all_to_all_tile_offset_bytes;
+            for (uint32_t i = 0; i < num_tiles_to_read; i++) {
+                uint32_t l1_write_addr_external = get_write_ptr(cb_external);
+                for(uint32_t block = 0; block < num_blocks; block++) {
+                    cb_reserve_back(cb_external, 1);
+                    uint64_t noc_addr_ex_par = remote_noc_addrs[block] | l1_read_addr_ex_par;
+                    noc_async_read_one_packet(noc_addr_ex_par, l1_write_addr_external, single_tile_size_bytes);
+                    l1_write_addr_external += single_tile_size_bytes;
 
-                noc_async_read_barrier();
-                cb_push_back(cb_ex_external, 1);
+                    noc_async_read_barrier();
+                    cb_push_back(cb_external, 1);
+                }
+                l1_read_addr_ex_par += single_tile_size_bytes;
             }
-            l1_read_addr_ex_par += single_tile_size_bytes;
-        }
 
-        // send result to other cores
-        uint32_t l1_write_addr_ex = get_write_ptr(cb_ex);
-        uint32_t l1_write_addr_ex_global = get_write_ptr(cb_ex_global);
-        cb_wait_front(cb_ex, num_tiles_to_read);
+            // send result to other cores
+            cb_wait_front(cb_ex, num_tiles_to_read);
+        }
 
         // sync with other workers
         noc_semaphore_set(reduce_sender_semaphore_addr_ptr, INVALID);
@@ -79,107 +115,14 @@ void kernel_main() {
         noc_semaphore_wait(reduce_sender_semaphore_addr_ptr, VALID);
         noc_semaphore_set(reduce_sender_semaphore_addr_ptr, INVALID);
 
-        for (uint32_t block = 0; block < num_all_to_all_workers; block++) {
-            uint32_t num_tiles = block == num_all_to_all_workers - 1 ? num_tiles_per_worker_last : num_tiles_per_worker;
-            cb_reserve_back(cb_ex_global, num_tiles);
-            noc_semaphore_wait(reduce_sender_semaphore_addr_ptr, block+1);
-            cb_push_back(cb_ex_global, num_tiles);
-
-        }
-
-    } else {
-        // wait for local data ready
-        cb_wait_front(cb_ex_partial, block_h);
-
-        // inc top core
-        noc_semaphore_set(reduce_sender_semaphore_addr_ptr, INVALID);
-        const uint64_t reduce_receiver_semaphore_noc_addr = get_noc_addr(noc_same_coord, 1, reduce_receiver_semaphore_addr);
-        noc_semaphore_inc(reduce_receiver_semaphore_noc_addr, 1);
-        noc_semaphore_wait(reduce_sender_semaphore_addr_ptr, VALID);
-
-        // sync with other workers
-        noc_semaphore_set(reduce_sender_semaphore_addr_ptr, INVALID);
-        noc_semaphore_inc(reduce_receiver_semaphore_noc_addr, 1);
-        noc_semaphore_wait(reduce_sender_semaphore_addr_ptr, VALID);
-        noc_semaphore_set(reduce_sender_semaphore_addr_ptr, INVALID);
-
-        for (uint32_t block = 0; block < num_all_to_all_workers; block++) {
+        for (uint32_t block = 0; block < num_all_to_all_workers; ++block) {
             uint32_t num_tiles = block == num_all_to_all_workers - 1 ? num_tiles_per_worker_last : num_tiles_per_worker;
             cb_reserve_back(cb_ex_global, num_tiles);
             noc_semaphore_wait(reduce_sender_semaphore_addr_ptr, block+1);
             cb_push_back(cb_ex_global, num_tiles);
         }
-
     };
-
-    // global reduce
-    if constexpr(is_all_to_all_worker) {
-        // wait for local data ready
-        cb_wait_front(cb_ex_partial2, block_h);
-        // inc semaphore of other cores
-        noc_semaphore_set(reduce_sender_semaphore_addr_ptr, INVALID);
-        const uint64_t reduce_receiver_semaphore_noc_addr = get_noc_addr(noc_same_coord, 1, reduce_receiver_semaphore_addr);
-        noc_semaphore_inc(reduce_receiver_semaphore_noc_addr, 1);
-        noc_semaphore_wait(reduce_sender_semaphore_addr_ptr, VALID);
-
-        // read data from other cores
-        uint32_t l1_read_addr_ex_par = get_read_ptr(cb_ex_partial2);
-        l1_read_addr_ex_par += all_to_all_tile_offset_bytes;
-        for (uint32_t i = 0; i < num_tiles_to_read; i++) {
-            uint32_t l1_write_addr_external = get_write_ptr(cb_ex_external2);
-            for(uint32_t block = 0; block < num_blocks; block++) {
-                cb_reserve_back(cb_ex_external2, 1);
-                uint64_t noc_addr_ex_par = get_noc_addr(noc_same_coord, noc_diff_coord[block], l1_read_addr_ex_par);
-                noc_async_read_one_packet(noc_addr_ex_par, l1_write_addr_external, single_tile_size_bytes);
-                l1_write_addr_external += single_tile_size_bytes;
-
-                noc_async_read_barrier();
-                cb_push_back(cb_ex_external2, 1);
-            }
-            l1_read_addr_ex_par += single_tile_size_bytes;
-        }
-
-        // send result to other cores
-        uint32_t l1_write_addr_ex = get_write_ptr(cb_ex2pe);
-        uint32_t l1_write_addr_ex_global = get_write_ptr(cb_ex_global);
-        cb_wait_front(cb_ex2pe, num_tiles_to_read);
-
-        // sync with other workers
-        noc_semaphore_set(reduce_sender_semaphore_addr_ptr, INVALID);
-        noc_semaphore_inc(reduce_receiver_semaphore_noc_addr, 1);
-        noc_semaphore_wait(reduce_sender_semaphore_addr_ptr, VALID);
-        noc_semaphore_set(reduce_sender_semaphore_addr_ptr, INVALID);
-
-        for (uint32_t block = 0; block < num_all_to_all_workers; block++) {
-            uint32_t num_tiles = block == num_all_to_all_workers - 1 ? num_tiles_per_worker_last : num_tiles_per_worker;
-            cb_reserve_back(cb_ex_global, num_tiles);
-            noc_semaphore_wait(reduce_sender_semaphore_addr_ptr, block+1);
-            cb_push_back(cb_ex_global, num_tiles);
-        }
-
-    } else {
-        // wait for local data ready
-        cb_wait_front(cb_ex_partial2, block_h);
-        // inc semaphore of other cores
-        noc_semaphore_set(reduce_sender_semaphore_addr_ptr, INVALID);
-        const uint64_t reduce_receiver_semaphore_noc_addr = get_noc_addr(noc_same_coord, 1, reduce_receiver_semaphore_addr);
-        noc_semaphore_inc(reduce_receiver_semaphore_noc_addr, 1);
-        noc_semaphore_wait(reduce_sender_semaphore_addr_ptr, VALID);
-
-        // // sync with other workers
-        noc_semaphore_set(reduce_sender_semaphore_addr_ptr, INVALID);
-        noc_semaphore_inc(reduce_receiver_semaphore_noc_addr, 1);
-        noc_semaphore_wait(reduce_sender_semaphore_addr_ptr, VALID);
-        noc_semaphore_set(reduce_sender_semaphore_addr_ptr, INVALID);
-
-        for (uint32_t block = 0; block < num_all_to_all_workers; block++) {
-            uint32_t num_tiles = block == num_all_to_all_workers - 1 ? num_tiles_per_worker_last : num_tiles_per_worker;
-            cb_reserve_back(cb_ex_global, num_tiles);
-            noc_semaphore_wait(reduce_sender_semaphore_addr_ptr, block+1);
-            cb_push_back(cb_ex_global, num_tiles);
-        }
-
-    };
-
+    global_reduce_receiver(cb_ex_partial, cb_ex_external, cb_ex, cb_ex_global);
+    global_reduce_receiver(cb_ex_partial2, cb_ex_external2, cb_ex2pe, cb_ex_global);
 
 }
