+++ tt_eager/tt_dnn/op_library/reduce/kernels/dataflow/reader_unary_reduce_interleaved_start_id.cpp
@@ -4,6 +4,7 @@
 
 #include <stdint.h>
 #include "dataflow_api.h"
+#include "tt_eager/tt_dnn/kernels/dataflow/generate_reduce_scaler.hpp"
 
 
 void kernel_main() {
@@ -14,29 +15,7 @@ void kernel_main() {
     constexpr uint32_t scaler = get_compile_time_arg_val(1);
 
     constexpr uint32_t cb_id_in2 = 2;
-    cb_reserve_back(cb_id_in2, 1);
-    constexpr uint32_t num_zeros_reads = 2048 / MEM_ZEROS_SIZE;
-    uint64_t zeros_noc_addr = get_noc_addr(MEM_ZEROS_BASE);
-    uint32_t write_addr = get_write_ptr(cb_id_in2);
-    // Fill tile with zeros
-    for (uint32_t i = 0; i < num_zeros_reads; ++i) {
-        noc_async_read(zeros_noc_addr, write_addr, MEM_ZEROS_SIZE);
-        write_addr += MEM_ZEROS_SIZE;
-    }
-    noc_async_read_barrier();
-    if constexpr (scaler != 0) {
-        volatile tt_l1_ptr uint32_t* ptr = reinterpret_cast<volatile tt_l1_ptr uint32_t*>(get_write_ptr(cb_id_in2));
-        uint32_t idx = 0;
-        for (uint32_t k = 0; k < 4; ++k) {
-            uint32_t curr_idx = idx;
-            for (uint32_t j = 0; j < 8; ++j) {
-                ptr[curr_idx] = scaler;
-                curr_idx++;
-            }
-            idx += 128;
-        }
-    }
-    cb_push_back(cb_id_in2, 1);
+    generate_reduce_scaler(cb_id_in2, scaler);
 
     constexpr uint32_t cb_id_in0 = 0;
