+++ tt_eager/tt_dnn/op_library/softmax/kernels/dataflow/reader_unary_interleaved_sm.cpp
@@ -3,45 +3,8 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #include "dataflow_api.h"
-
-FORCE_INLINE void generate_bcast_scaler() {
-    constexpr uint32_t cb_in_2 = 2;
-    uint32_t scaler = get_arg_val<uint32_t>(10);
-    //DPRINT << "basic Scaler = " << F32(u.f) << ENDL();
-    cb_reserve_back(cb_in_2, 1);
-    constexpr uint32_t num_zeros_reads = 2048 / MEM_ZEROS_SIZE;
-    uint64_t zeros_noc_addr = get_noc_addr(MEM_ZEROS_BASE);
-    uint32_t write_addr = get_write_ptr(cb_in_2);
-    // Fill tile with zeros
-    for (uint32_t i = 0; i < num_zeros_reads; ++i) {
-        noc_async_read(zeros_noc_addr, write_addr, MEM_ZEROS_SIZE);
-        write_addr += MEM_ZEROS_SIZE;
-    }
-    noc_async_read_barrier();
-    volatile tt_l1_ptr uint32_t* ptr = reinterpret_cast<volatile tt_l1_ptr uint32_t*>(get_write_ptr(cb_in_2));
-    uint32_t idx = 0;
-    for (uint32_t k = 0; k < 4; ++k) {
-        uint32_t curr_idx = idx;
-        for (uint32_t j = 0; j < 8; ++j) {
-            ptr[curr_idx] = scaler;
-            curr_idx++;
-        }
-        idx += 128;
-    }
-    cb_push_back(cb_in_2, 1);
-}
-
-// HW-bcast scale for fused scale-attn-softmax
-FORCE_INLINE void generate_inv_sqrt_hw_bcast_tile() {
-    constexpr uint32_t scale_cb_id = 3;
-    uint32_t u = get_arg_val<uint32_t>(2);
-    cb_reserve_back(scale_cb_id, 1);
-    auto ptr = reinterpret_cast<uint16_t*>(get_write_ptr(scale_cb_id));
-    // for (int j = 0; j < 1024; j++)
-    //     ptr[j] = uint16_t(0);
-    ptr[0] = u>>16;
-    cb_push_back(scale_cb_id, 1);
-}
+#include "tt_eager/tt_dnn/kernels/dataflow/generate_reduce_scaler.hpp"
+#include "tt_eager/tt_dnn/kernels/dataflow/generate_bcast_scalar.hpp"
 
 void kernel_main() {
 
@@ -88,7 +51,9 @@ void kernel_main() {
     uint32_t ht = start_ht;
     uint32_t mask_id = start_mask_id;
     bool read_mask = true;
-    generate_inv_sqrt_hw_bcast_tile();
+    constexpr auto cb_fused_scale = tt::CB::c_in3;
+    const uint32_t pre_scale = get_arg_val<uint32_t>(2);
+    generate_bcast_unary_scalar(cb_fused_scale, pre_scale);
     #endif
 
     const InterleavedAddrGenFast<src0_is_dram> src_a = {
@@ -99,7 +64,11 @@ void kernel_main() {
 
 
     // TODO(AP): cleanup, probably with named args/param pack/reflection.
-    generate_bcast_scaler();
+    {
+        constexpr uint32_t cb_in_2 = 2;
+        const uint32_t reduce_scaler = get_arg_val<uint32_t>(10);
+        generate_reduce_scaler(cb_in_2, reduce_scaler);
+    }
 
     // read a ublock of tiles from src to CB, and then push the ublock to unpacker
     uint32_t i_tile = 0;
