+++ tt_eager/tt_dnn/op_library/softmax/kernels/dataflow/reader_unary_sharded_sm.cpp
@@ -3,23 +3,8 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #include "dataflow_api.h"
-
-FORCE_INLINE void generate_bcast_scaler() {
-    constexpr auto cb_bcast_scaler = tt::CB::c_in1;
-    uint32_t scaler = get_arg_val<uint32_t>(0);
-    cb_reserve_back(cb_bcast_scaler, 1);
-    auto ptr = reinterpret_cast<volatile tt_l1_ptr uint32_t*>(get_write_ptr(cb_bcast_scaler));
-    uint32_t idx = 0;
-    for (uint32_t k = 0; k < 4; ++k) {
-        uint32_t curr_idx = idx;
-        for (uint32_t j = 0; j < 8; ++j) {
-            ptr[curr_idx] = scaler;
-            curr_idx++;
-        }
-        idx += 128;
-    }
-    cb_push_back(cb_bcast_scaler, 1);
-}
+#include "tt_eager/tt_dnn/kernels/dataflow/generate_reduce_scaler.hpp"
+#include "tt_eager/tt_dnn/kernels/dataflow/generate_bcast_scalar.hpp"
 
 // HW-bcast scale for fused scale-attn-softmax
 FORCE_INLINE void generate_inv_sqrt_hw_bcast_tile() {
@@ -32,6 +17,9 @@ FORCE_INLINE void generate_inv_sqrt_hw_bcast_tile() {
 }
 
 void kernel_main() {
+
+    constexpr uint32_t cb_reduce_scaler = tt::CB::c_in1;
+    const uint32_t reduce_scaler = get_arg_val<uint32_t>(0);
 
     #if FUSED_SCALE_MASK
     constexpr uint32_t block_wt = get_compile_time_arg_val(0);
@@ -50,7 +38,9 @@ void kernel_main() {
         .data_format = mask_data_format
     };
 
-    generate_inv_sqrt_hw_bcast_tile();
+    constexpr auto cb_fused_scale = tt::CB::c_in2;
+    const uint32_t pre_scale = get_arg_val<uint32_t>(1);
+    generate_bcast_unary_scalar(cb_fused_scale, pre_scale);
 
     #if CAUSAL_MASK
     uint32_t fused_head = get_compile_time_arg_val(4);
@@ -68,7 +58,7 @@ void kernel_main() {
             cb_push_back(cb_attn, block_wt);
 
             if (f == 0 && h == 0) {
-                generate_bcast_scaler();
+                generate_reduce_scaler(cb_reduce_scaler, reduce_scaler);
             }
         }
     }
@@ -83,10 +73,10 @@ void kernel_main() {
     noc_async_read_barrier();
     cb_push_back(cb_attn, block_wt);
 
-    generate_bcast_scaler();
+    generate_reduce_scaler(cb_reduce_scaler, reduce_scaler);
     #endif
 
     #else
-    generate_bcast_scaler();
+    generate_reduce_scaler(cb_reduce_scaler, reduce_scaler);
     #endif
 }
