+++ tt_eager/tt_lib/csrc/tt_lib_bindings_tensor.cpp
@@ -709,7 +709,7 @@ void TensorModule(py::module &m_tensor) {
 
     m_tensor.def(
         "num_cores_to_corerange_set",
-        &num_cores_to_corerange_set,
+        py::overload_cast<const uint32_t, const CoreCoord, const bool>(&num_cores_to_corerange_set),
         R"doc(
             Create a CoreRangeSet containing the specified number of cores
         )doc"
